fix norm

+++ ft_strlcat.c
@@ -6,19 +6,18 @@
 /*   By: almelo <almelo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/10 15:10:42 by almelo            #+#    #+#             */
-/*   Updated: 2022/05/19 15:39:12 by almelo           ###   ########.fr       */
+/*   Updated: 2022/05/21 09:41:42 by almelo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
-#include <stdio.h>
 
 size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 {
-	size_t	dst_len;
-	size_t	i;
+	size_t		dst_len;
+	unsigned int	i;
 
-	dst_len = ft_strlen(dst);	
+	dst_len = ft_strlen(dst);
 	i = 0;
 	while (i < (dstsize - dst_len - 1) && *(src + i))
 	{
